ENH/RF: Refactor of code component and coder with base class The code component needed better commenting and smart indenting for JS code. The refactor brings common attributes and methods for Coder and code component and puts them into a new base class. The toggleCommentLines and smartIdentThisLine methods support code type from the code component, but will always be set to python for Coder.
